+++ src/functions/quota-collector/main.py
@@ -12,6 +12,9 @@ from modules.qmchecks.ec2.ec2 import get_current_quotastatus_ec2
 from modules.qmchecks.ebs import get_current_quotastatus_ebs
 from modules.qmchecks.ec2_ipam import get_current_quotastatus_ec2_ipam
 from modules.qmchecks.customer_profiles import get_current_quotastatus_customer_profiles
+from modules.qmchecks.drs import get_current_quotastatus_drs
+from modules.qmchecks.schemas import get_current_quotastatus_schemas
+from modules.qmchecks.launchwizard import get_current_quotastatus_launchwizard
 from modules.qmchecks.vpc.vpc import get_current_quotastatus_vpc
 from modules.qmchecks.lambda_checks.lambda_checks import get_current_quotastatus_lambda
 from modules.qmchecks.account_services import get_current_quotastatus_account_services
@@ -227,6 +230,9 @@ def lambda_handler(event, context):
     entries.extend(get_current_quotastatus_ebs(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_ec2_ipam(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_customer_profiles(ctx=ctx, skip=official))
+    entries.extend(get_current_quotastatus_drs(ctx=ctx, skip=official))
+    entries.extend(get_current_quotastatus_schemas(ctx=ctx, skip=official))
+    entries.extend(get_current_quotastatus_launchwizard(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_account_services(ctx, skip=official))
     entries.extend(get_current_quotastatus_elb(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_stepfunctions(ctx=ctx, skip=official))